requirements.txt
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0
rich>=13.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
apscheduler>=3.10.4
